:hammer: refector: hooks 분리
